requirements.txt
mcp
langchain
langchain-mcp-adapters
langchain-huggingface
huggingface_hub
python-pptx  
requests